refactor: refactors connector structure for improved modularity
Moves and renames connector files to a dedicated connectors directory,
streamlining imports and clarifying project organization. Enhances
maintainability by separating integration logic from document types.
